requirements.txt
google-api-python-client
google-auth-httplib2
google-auth-oauthlib
jinja2
python-dotenv
schedule
mcp
pandas
